requirements.txt
asgiref==3.5.2
Django==4.1.3
psycopg2-binary
python-dotenv
whitenoise
django-extensions==3.2.1
django-import-export==3.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pyparsing==3.0.9
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
psycopg2==2.8.6
docutils==0.19
